model, generator, interfaces: isolating ISA-agnostic parts from x86
Includes:
- Target Description
- Taint Tracker
- BPAS contract
