+++ src/x86/x86_model.py
@@ -4,17 +4,14 @@ File: x86-specific model implementation
 Copyright (C) Microsoft Corporation
 SPDX-License-Identifier: MIT
 """
-import re
-import copy
 import numpy as np
-from typing import List, Tuple, Dict, Optional, Set
+from typing import Tuple, Dict
 
 import unicorn.x86_const as ucc
 from unicorn import Uc, UC_MEM_WRITE, UC_ARCH_X86, UC_MODE_64
 
-from interfaces import InputTaint, Instruction, RegisterOperand, FlagsOperand, \
-    MemoryOperand, Input
-from model import UnicornModel, UnicornSpec, UnicornSeq, TaintTrackerInterface
+from interfaces import Input
+from model import UnicornModel, UnicornSpec, UnicornSeq, UnicornBpas, BaseTaintTracker
 from x86.x86_target_desc import X86UnicornTargetDesc, X86TargetDesc
 
 FLAGS_CF = 0b000000000001
@@ -220,42 +217,8 @@ class X86UnicornCond(X86UnicornSpec):
         pass  # cond does not need to speculate mem accesses
 
 
-class X86UnicornBpas(X86UnicornSpec):
-
-    @staticmethod
-    def speculate_mem_access(emulator, access, address, size, value, model):
-        """
-        Since Unicorn does not have post-instruction hooks,
-        I have to implement it in a dirty way:
-        Save the information about the store here, but execute all the
-        contract logic in a hook before the next instruction (see trace_instruction)
-        """
-        if access == UC_MEM_WRITE:
-            rip = emulator.reg_read(ucc.UC_X86_REG_RIP)
-            opcode = emulator.mem_read(rip, 1)[0]
-            if opcode not in [0xE8, 0xFF, 0x9A]:  # ignore CALL instructions
-                model.previous_store = (address, size, emulator.mem_read(address, size), value)
-
-    @staticmethod
-    def speculate_instruction(emulator: Uc, address, _, model) -> None:
-        # reached max spec. window? skip
-        if len(model.checkpoints) >= model.nesting:
-            return
-
-        if model.previous_store[0]:
-            store_addr = model.previous_store[0]
-            old_value = bytes(model.previous_store[2])
-            new_is_signed = model.previous_store[3] < 0
-            new_value = (model.previous_store[3]). \
-                to_bytes(model.previous_store[1], byteorder='little', signed=new_is_signed)
-
-            # store a checkpoint
-            model.checkpoint(emulator, address)
-
-            # cancel the previous store but preserve its value
-            emulator.mem_write(store_addr, old_value)
-            model.store_logs[-1].append((store_addr, new_value))
-        model.previous_store = (0, 0, 0, 0)
+class X86UnicornBpas(UnicornBpas, X86UnicornModel):
+    pass
 
 
 class X86UnicornNull(X86UnicornSpec):
@@ -294,7 +257,7 @@ class X86UnicornNull(X86UnicornSpec):
 class X86UnicornCondBpas(X86UnicornSpec):
 
     @staticmethod
-    def speculate_mem_access(emulator, access, address, size, value, model):
+    def speculate_mem_access(emulator, access, address, size, value, model: UnicornSpec):
         X86UnicornBpas.speculate_mem_access(emulator, access, address, size, value, model)
 
     @staticmethod
@@ -306,194 +269,16 @@ class X86UnicornCondBpas(X86UnicornSpec):
 # ==================================================================================================
 # Taint tracker
 # ==================================================================================================
-class X86TaintTracker(TaintTrackerInterface):
-    strict_undefined: bool = True
-    _instruction: Optional[Instruction] = None
-    sandbox_base: int = 0
-
-    src_regs: List[str]
-    dest_regs: List[str]
-    reg_dependencies: Dict[str, Set]
-
-    src_flags: List[str]
-    dest_flags: List[str]
-    flag_dependencies: Dict[str, Set]
-
-    src_mems: List[str]
-    dest_mems: List[str]
-    mem_dependencies: Dict[str, Set]
-
-    mem_address_regs: List[str]
-
-    tainted_labels: Set[str]
-    pending_taint: List[str]
-
+class X86TaintTracker(BaseTaintTracker):
+    # ISA-specific fields
     _registers = [
         ucc.UC_X86_REG_RAX, ucc.UC_X86_REG_RBX, ucc.UC_X86_REG_RCX, ucc.UC_X86_REG_RDX,
         ucc.UC_X86_REG_RSI, ucc.UC_X86_REG_RDI, ucc.UC_X86_REG_EFLAGS
     ]
 
     def __init__(self, initial_observations, sandbox_base=0):
-        self.initial_observations = initial_observations
-        self.sandbox_base = sandbox_base
-        self.flag_dependencies = {}
-        self.reg_dependencies = {}
-        self.mem_dependencies = {}
-        self.tainted_labels = set(self.initial_observations)
-        self.checkpoints = []
-        self.target_desc = X86UnicornTargetDesc()
+        super().__init__(initial_observations, sandbox_base=sandbox_base)
 
-    def start_instruction(self, instruction):
-        """ Collect source and target registers/flags """
-        if self._instruction:
-            self._finalize_instruction()  # finalize the previous instruction
-
-        self._instruction = instruction
-        self.src_regs = []
-        self.src_flags = []
-        self.src_mems = []
-        self.dest_regs = []
-        self.dest_flags = []
-        self.dest_mems = []
-        self.pending_taint = []
-        self.mem_address_regs = []
-
-        for op in instruction.get_all_operands():
-            if isinstance(op, RegisterOperand):
-                value = X86TargetDesc.gpr_normalized[op.value]
-                if op.src:
-                    self.src_regs.append(value)
-                if op.dest:
-                    self.dest_regs.append(value)
-            elif isinstance(op, FlagsOperand):
-                self.src_flags = op.get_read_flags()
-                if self.strict_undefined:
-                    self.src_flags.extend(op.get_undef_flags())
-                self.dest_flags = op.get_write_flags()
-            elif isinstance(op, MemoryOperand):
-                for sub_op in re.split(r'\+|-|\*| ', op.value):
-                    if sub_op and sub_op in X86TargetDesc.gpr_normalized:
-                        self.mem_address_regs.append(X86TargetDesc.gpr_normalized[sub_op])
-
-    def _finalize_instruction(self):
-        """Propagate dependencies from source operands to destinations """
-
-        # Compute source label
-        src_labels = set()
-        for reg in self.src_regs:
-            src_labels.update(self.reg_dependencies.get(reg, {reg}))
-        for flag in self.src_flags:
-            src_labels.update(self.flag_dependencies.get(flag, {flag}))
-        for addr in self.src_mems:
-            src_labels.update(self.mem_dependencies.get(addr, {addr}))
-
-        # print(src_labels)
-
-        # Propagate label to all targets
-        uniq_labels = src_labels
-        for reg in self.dest_regs:
-            if reg in self.reg_dependencies:
-                self.reg_dependencies[reg].update(uniq_labels)
-            else:
-                self.reg_dependencies[reg] = copy.copy(uniq_labels)
-                self.reg_dependencies[reg].add(reg)
-
-        for flg in self.dest_flags:
-            if flg in self.flag_dependencies:
-                self.flag_dependencies[flg].update(uniq_labels)
-            else:
-                self.flag_dependencies[flg] = copy.copy(uniq_labels)
-                self.flag_dependencies[flg].add(flg)
-
-        for mem in self.dest_mems:
-            if mem in self.mem_dependencies:
-                self.mem_dependencies[mem].update(uniq_labels)
-            else:
-                self.mem_dependencies[mem] = copy.copy(uniq_labels)
-                self.mem_dependencies[mem].add(mem)
-
-        # Update taints
-        for label in self.pending_taint:
-            if label.startswith("0x"):
-                self.tainted_labels.update(self.mem_dependencies.get(label, {label}))
-            else:
-                self.tainted_labels.update(self.reg_dependencies.get(label, {label}))
-
-        self._instruction = None
-
-    def track_memory_access(self, address: int, size: int, is_write: bool):
-        """ Tracking concrete memory accesses """
-        # mask the address - we taint at the granularity of 8 bytes
-        address -= self.sandbox_base
-        masked_start_addr = address & 0xffff_ffff_ffff_fff8
-        end_addr = address + (size - 1)
-        masked_end_addr = end_addr & 0xffff_ffff_ffff_fff8
-
-        # add all addresses to tracking
-        track_list = self.dest_mems if is_write else self.src_mems
-        for i in range(masked_start_addr, masked_end_addr + 1, 8):
-            track_list.append(hex(i))
-
-    def taint_pc(self):
-        if self._instruction and self._instruction.control_flow:
-            self.pending_taint.append("RIP")
-
-    def taint_memory_access_address(self):
-        for reg in self.mem_address_regs:
-            self.pending_taint.append(reg)
-
-    def taint_memory_load(self):
-        for addr in self.src_mems:
-            self.pending_taint.append(addr)
-
-    def taint_memory_store(self):
-        for addr in self.dest_mems:
-            self.pending_taint.append(addr)
-
-    def checkpoint(self):
-        if self._instruction:
-            self._finalize_instruction()
-        self.checkpoints.append(
-            (copy.deepcopy(self.flag_dependencies), copy.deepcopy(self.reg_dependencies),
-             copy.deepcopy(self.mem_dependencies)))
-
-    def rollback(self):
-        assert self.checkpoints, "There are no more checkpoints"
-        if self._instruction:
-            self._finalize_instruction()
-        t = self.checkpoints.pop()
-        self.flag_dependencies = copy.deepcopy(t[0])
-        self.reg_dependencies = copy.deepcopy(t[1])
-        self.mem_dependencies = copy.deepcopy(t[2])
-
-    def get_taint(self) -> InputTaint:
-        if self._instruction:
-            self._finalize_instruction()
-
-        taint = InputTaint()
-        tainted_positions = []
-        register_start = taint.register_start
-
-        for label in self.tainted_labels:
-            input_offset = -1  # the location of the label within the Input array
-            if label.startswith('0x'):
-                # memory address
-                # we taint the 64-bits block that contains the address
-                input_offset = (int(label, 16)) // 8
-            else:
-                reg = self.target_desc.reg_decode[label]
-                if reg in self._registers:
-                    input_offset = register_start + \
-                          self._registers.index(self.target_desc.reg_decode[label])
-            if input_offset >= 0:
-                tainted_positions.append(input_offset)
-
-        tainted_positions = list(dict.fromkeys(tainted_positions))
-        tainted_positions.sort()
-        for i in range(taint.size):
-            if i in tainted_positions:
-                taint[i] = True
-            else:
-                taint[i] = False
-
-        return taint
+        # ISA-specific field setup
+        self.target_desc = X86TargetDesc()
+        self.unicorn_target_desc = X86UnicornTargetDesc()
